lib: extract Flatten, Equivalent, Contains{,All}, and RoleName from core lib to extras/resources
